Trim old comments
