+++ rmf_fleet_adapter/rmf_rxcpp/include/rmf_rxcpp/RxJobs.hpp
@@ -31,12 +31,6 @@ inline auto make_job(const std::shared_ptr<Action>& action)
   return detail::make_observable<T>(action);
 }
 
-//template<typename T, typename F>
-//inline auto make_leaky_job(const F& f)
-//{
-//  return detail::make_observable<T>(std::make_shared<F>(f));
-//}
-
 template<typename Job0, typename... Jobs>
 inline auto merge_jobs(const Job0& o0, Jobs&& ... os)
 {
